moved define wrapper to own file
