+++ qcengine/programs/turbomole/methods.py
@@ -1,0 +1,2 @@
+
+ricc2_methods = "rimp2 rimp3 rimp4 ricc2 riccsd riccsd(t)".split()
